Added a priority attribute to talks
Since we're starting to add low-priority talks (i.e. talks that are
mostly proposed to fill in a short line-up, if need be), having a
clear wat to point them out could be useful. And while we're at it, why
not an the option of marking a talk as "high priority" so that room
could be made for it (e.g. for an special guest).

Note that this attribute is not user-facing – only admins should have
the ability to decide that a talk as is low- or high-priority. By
default, the priority of a talk is "normal", and this level should be
kept most of time.